Rename args to args_parser
I want to save a globally-accessible variable called `args` to `args.py`. This may get confusing, so renaming `args.py` to `args_parser.py` should help.

For example, now every module can access the command line args that were parsed:

    from app_deployer import args_parser

    args = args_parser.args
